More UML Changes...
Added a deleteProfile() and getPendingStatus(arrDep:ArriveDepart):bool
methods to the Person class.
Changed the getter/setter for driver in Group class. setDriver() now
requires a param for the day of week as well as the person we are
setting as the driver. getDriver() now also requires a param for the day
of the week.